Move AbbreviationMatcher into CommandLine.java
